feat: add `changelog-host` input to `action.yml` (#948)
This was originally part of
[3.3.0](https://github.com/google-github-actions/release-please-action/blob/main/CHANGELOG.md#330-2022-08-10)
(and fixed in
[3.4.1](https://github.com/google-github-actions/release-please-action/blob/main/CHANGELOG.md#341-2022-08-24)
to default to `github.server_url`). However, it was not part of 4.0.0.

Without this input, you have to provide both config and manifest files
in order to set the `changelog-host` option.

With this input, you can simply provide `release-type` and
`changelog-host` without the need for config and manifest files. Also,
since it defaults to `github.server_url`, most users will not even have
to set it.

Signed-off-by: Paavo Bennett <paavobennettis@gmail.com>
Co-authored-by: Jeff Ching <chingor@google.com>